Read settings.ini in jupyter hooks
Jupyter hooks are run from Jupyter's cwd (can be specified or else directory Jupyter was launched from)

When this isn't the same directory as the nbdev repos root directory the settings file is not found.
Because some settings have default values e.g. allowed_metadata_keys the hook doesn't fail when the settings file is missing.

Instead we look at the path of the notebook given to the hook and attempt to find the settings file by
traversing directories upward until we find it or hit root '/'.