exported modules can scrub_magics via settings.ini
This scrubs the cell magic lines out in the generated code. In some
sense, it is a follow-up to fastai/nbdev#905.

Would y'all be interested in a PR that adds a new option to the
settings.ini config? It would scrub any `%%magics` from the exported
modules. Fairly similar to how `black_formatting` works today.

---

The specific use-case is to simplify doing exploratory programming in
PySpark. We can now use nbdev notebooks to generate valid [pyspark][0]
code modules. At the same time, the Python-kernel notebook can continue
to use the [`%%spark`][1] magic. We can use a local notebook to interact
with a remote cluster running Livy.

[0]: https://github.com/apache/spark/tree/v3.1.1-rc3/examples/src/main/python
[1]: https://github.com/jupyter-incubator/sparkmagic

---

Please note that the implementation is just copying
`processors.clean_magics`. Not sure how much of an issue that is for
this repository. Should we add a code cell like below above the
scrub_magics method?

```
\#|hide
\# FIXME: reusing processors.clean_magics runs into a circular import
\# from .processors import clean_magics
```