Clarify in tutorial about gh-pages branch creation
I got stuck at this point of the nbdev End-to-End Walkthrough because I
didn't see a gh-pages branch and was unsure if I should create it or if
the tutorial was outdated.
Fortunately I decided to refresh the GitHub Pages settings page after a
few minutes and saw it created.